requirements.txt
pymorphy3==2.0.3
python-dotenv==1.0.1
python-telegram-bot==22.0
